Update Skiko to 0.7.84, fix failed tests after updating coroutines (#871)
- Update Skiko to 0.7.84
([Diff](https://github.com/JetBrains/skiko/compare/v0.7.80...v0.7.84))
- Because it updates Coroutines to 1.7.3, now we have failing tests,
they are migrated to the new coroutine version

Having 1.7.3 in dependencies is okay, as it depends on [Kotlin
1.8.20](https://github.com/Kotlin/kotlinx.coroutines/blob/master/CHANGES_UP_TO_1.7.md#version-170)
(Compose 1.5 depends on 1.8.21). Besides that, [Jetpack Compose 1.6
already depends on
1.7.1](https://github.com/androidx/androidx/blob/e502da69dcec7dc8662241de58f34399e7b0abe8/gradle/libs.versions.toml#L47)
