+++ compose/foundation/foundation/src/test/kotlin/androidx/compose/foundation/gestures/BringIntoViewRequestPriorityQueueTest.kt
@@ -24,6 +24,7 @@ import com.google.common.truth.Truth.assertThat
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
 import kotlinx.coroutines.CancellableContinuation
+import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.CoroutineStart
@@ -191,7 +192,7 @@ class BringIntoViewRequestPriorityQueueTest {
         queue.enqueue(smallRequest)
         queue.enqueue(mediumRequest)
         queue.enqueue(largeRequest)
-        val cause = RuntimeException("Oops")
+        val cause = CancellationException("Oops")
 
         queue.cancelAndRemoveAll(cause)
