+++ gradle/libs.versions.toml
@@ -54,7 +54,7 @@ moshi = "1.13.0"
 protobuf = "3.21.8"
 paparazzi = "1.0.0"
 paparazziNative = "2022.1.1-canary-f5f9f71"
-skiko = "0.7.80"
+skiko = "0.7.84"
 sqldelight = "1.3.0"
 retrofit = "2.7.2"
 wire = "4.5.1"
